reorganize commsSlots files
